@@ -17,86 +16,63 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/testuser/material.git"
+    "url": "git+https://github.com/prescottprue/material.git"
   },
   "engines": {
-    "node": ">=6.11.5"
+    "node": ">=8"
   },
-  "author": "testuser (https://github.com/testuser)",
+  "author": "prescottprue (https://github.com/prescottprue)",
   "license": "MIT",
   "dependencies": {
-    "@material-ui/core": "1.0.0",
-    "@material-ui/icons": "1.0.0",
-    "firebase": "^5.0.3",
-    "lodash": "^4.17.10",
-    "material-ui": "next",
-    "normalize.css": "^7.0.0",
-    "object-assign": "^4.1.1",
-    "promise": "^7.1.1",
+    "@material-ui/core": "^3.5.1",
+    "@material-ui/icons": "^3.0.1",
+    "firebase": "^5.5.8",
+    "history": "^4.7.2",
+    "lodash": "^4.17.11",
     "prop-types": "^15.6.1",
-    "react": "^16.3.0",
-    "react-dom": "^16.3.0",
-    "react-google-button": "^0.3.0",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-scripts": "2.1.1",
+    "react-google-button": "^0.5.3",
+    "react-loadable": "^5.5.0",
     "react-redux": "^5.0.7",
-    "react-redux-firebase": "*",
-    "react-router": "^3.0.0",
-    "recompose": "^0.27.1",
-    "redbox-react": "^1.3.6",
+    "react-redux-firebase": "^2.1.9",
+    "react-router-dom": "^4.3.1",
+    "recompose": "^0.30.0",
     "redux": "^4.0.0",
-    "redux-auth-wrapper": "^1.0.0",
-    "redux-form": "^7.3.0",
+    "redux-auth-wrapper": "^2.0.3",
+    "redux-form": "^7.4.0",
     "redux-form-material-ui": "next",
-    "redux-thunk": "^2.2.0",
-    "whatwg-fetch": "^2.0.3"
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.0.0",
-    "babel-plugin-lodash": "^3.2.11",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-export-extensions": "^6.22.0",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-env": "^1.4.0",
-    "babel-preset-react": "^6.24.1",
-    "babel-runtime": "^6.20.0",
-    "chalk": "^1.1.3",
-    "compression": "^1.6.2",
-    "connect-history-api-fallback": "^1.3.0",
-    "cross-env": "^5.0.0",
-    "css-loader": "^0.28.1",
-    "eslint": "^3.19.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "eslint": "5.6.0",
     "eslint-config-prettier": "^2.3.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-config-standard-react": "^5.0.0",
-    "eslint-plugin-babel": "^4.1.1",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-node": "^4.2.2",
+    "eslint-config-react-app": "^3.0.4",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jsx-a11y": "6.1.2",
     "eslint-plugin-prettier": "^2.1.2",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-react": "^7.0.1",
-    "eslint-plugin-standard": "^3.0.1",
-    "express": "^4.14.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "figures": "^2.0.0",
-    "file-loader": "^0.11.1",
-    "fs-extra": "^3.0.1",
-    "html-webpack-plugin": "^2.29.0",
+    "eslint-plugin-react": "7.11.1",
     "husky": "^0.14.3",
-    "ip": "^1.1.5",
-    "node-sass": "^4.5.3",
-    "prettier": "^1.12.1",
-    "rimraf": "^2.6.1",
-    "sass-loader": "^6.0.5",
-    "style-loader": "^0.17.0",
-    "url-loader": "^0.5.8",
-    "webpack": "^3.6.0",
-    "webpack-dashboard": "^0.4.0",
-    "webpack-dev-middleware": "^1.11.0",
-    "webpack-hot-middleware": "^2.13.2",
-    "yargs": "^8.0.1"
+    "node-sass": "^4.10.0",
+    "prettier": "^1.14.3"
+  },
+  "eslintConfig": {
+    "extends": "./eslintrc.json"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ],
+  "babel": {
+    "presets": [
+      "react-app"
+    ],
+    "plugins": [
+      "@babel/plugin-syntax-dynamic-import"
+    ]
   }
 }
