@@ -1,22 +1,10 @@
-export default () => ({
-  path: '*',
-  /*  Async getComponent is only invoked when route matches   */
-  getComponent(nextState, cb) {
-    /*  Webpack - use 'require.ensure' to create a split point
-        and embed an async module loader (jsonp) when bundling   */
-    require.ensure(
-      [],
-      require => {
-        /*  Webpack - use require callback to define
-          dependencies for bundling   */
-        const NotFound = require('./NotFound').default
+import Loadable from 'react-loadable'
+import LoadingSpinner from 'components/LoadingSpinner'
 
-        /*  Return getComponent   */
-        cb(null, NotFound)
-
-        /* Webpack named bundle   */
-      },
-      'NotFound'
-    )
-  }
-})
+export default {
+  component: Loadable({
+    loader: () =>
+      import(/* webpackChunkName: 'NotFound' */ './components/NotFoundPage'),
+    loading: LoadingSpinner
+  })
+}
