@@ -1,22 +1,9 @@
-export default store => ({
-  path: ':projectname',
-  /*  Async getComponent is only invoked when route matches   */
-  getComponent(nextState, cb) {
-    /*  Webpack - use 'require.ensure' to create a split point
-        and embed an async module loader (jsonp) when bundling   */
-    require.ensure(
-      [],
-      require => {
-        /*  Webpack - use require callback to define
-          dependencies for bundling   */
-        const Project = require('./components/ProjectPage').default
+import { Loadable } from 'utils/components'
 
-        /*  Return getComponent   */
-        cb(null, Project)
-
-        /* Webpack named bundle   */
-      },
-      'Project'
-    )
-  }
-})
+export default {
+  path: ':projectId',
+  component: Loadable({
+    loader: () =>
+      import(/* webpackChunkName: 'Project' */ './components/ProjectPage')
+  })
+}
