@@ -1,10 +1,9 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import CircularProgress from '@material-ui/core/CircularProgress'
-import classes from './LoadingSpinner.scss'
 
-export const LoadingSpinner = ({ size }) => (
-  <div className={classes.container}>
+export const LoadingSpinner = ({ size, classes }) => (
+  <div className={classes.root}>
     <div className={classes.progress}>
       <CircularProgress mode="indeterminate" size={size || 80} />
     </div>
@@ -12,6 +11,7 @@ export const LoadingSpinner = ({ size }) => (
 )
 
 LoadingSpinner.propTypes = {
+  classes: PropTypes.object.isRequired,
   size: PropTypes.number
 }
 
