@@ -1,3 +1,4 @@
 import LoadingSpinner from './LoadingSpinner'
+import enhance from './LoadingSpinner.enhancer'
 
-export default LoadingSpinner
+export default enhance(LoadingSpinner)
