@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { browserHistory, Router } from 'react-router'
+import { BrowserRouter as Router } from 'react-router-dom'
 import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
 import { Provider } from 'react-redux'
 import ThemeSettings from 'theme'
@@ -10,7 +10,7 @@ const theme = createMuiTheme(ThemeSettings)
 const App = ({ routes, store }) => (
   <MuiThemeProvider theme={theme}>
     <Provider store={store}>
-      <Router history={browserHistory}>{routes}</Router>
+      <Router>{routes}</Router>
     </Provider>
   </MuiThemeProvider>
 )
