@@ -1,8 +1,7 @@
 import React from 'react'
-import { Link } from 'react-router'
+import { Link } from 'react-router-dom'
 import Button from '@material-ui/core/Button'
-import { LOGIN_PATH, SIGNUP_PATH } from 'constants'
-import classes from './Navbar.scss'
+import { LOGIN_PATH, SIGNUP_PATH } from 'constants/paths'
 
 const buttonStyle = {
   color: 'white',
@@ -11,7 +10,7 @@ const buttonStyle = {
 }
 
 export const LoginMenu = () => (
-  <div className={classes.menu}>
+  <div>
     <Button style={buttonStyle} component={Link} to={SIGNUP_PATH}>
       Sign Up
     </Button>
