@@ -1,13 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'react-router'
+import { Link } from 'react-router-dom'
 import AppBar from '@material-ui/core/AppBar'
 import Toolbar from '@material-ui/core/Toolbar'
 import Typography from '@material-ui/core/Typography'
+import { LIST_PATH } from 'constants/paths'
 import AccountMenu from './AccountMenu'
 import LoginMenu from './LoginMenu'
-import { LIST_PATH } from 'constants'
-import classes from './Navbar.scss'
 
 export const Navbar = ({
   avatarUrl,
@@ -17,12 +16,13 @@ export const Navbar = ({
   handleLogout,
   closeAccountMenu,
   anchorEl,
-  handleMenu
+  handleMenu,
+  classes
 }) => (
   <AppBar position="static">
     <Toolbar>
       <Typography
-        type="title"
+        variant="h6"
         color="inherit"
         className={classes.flex}
         component={Link}
@@ -47,6 +47,7 @@ export const Navbar = ({
 )
 
 Navbar.propTypes = {
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
   displayName: PropTypes.string, // from enhancer (flattenProps - profile)
   avatarUrl: PropTypes.string, // from enhancer (flattenProps - profile)
   authExists: PropTypes.bool, // from enhancer (withProps - auth)
