@@ -7,14 +7,14 @@ import {
 const defaultDismissTime = 2500 // 2.5 seconds
 
 /**
- * @description Publish a notification. if `dismissAfter` is set,
- * the notification will be auto dismissed after the given period.
+ * Publish a notification. if `dismissAfter` is set, the notification will be
+ * auto dismissed after the given period.
  * @param {Object} notif - Object containing
  * @param {Object} notif.kind - Kinda of notification (success, warning, failure)
  * @param {Object} notif.message - Notification message
  * @param {Object} notif.dismissAfter - Time after which to dismiss notification (default time set in constants)
  */
-export const showNotification = notif => {
+export function showNotification(notif) {
   const payload = Object.assign({}, notif)
   // Set default id to now if none provided
   if (!payload.id) {
@@ -32,22 +32,36 @@ export const showNotification = notif => {
   }
 }
 
-export const showSuccess = message =>
-  showNotification({ type: 'success', message })
-
-export const showError = message =>
-  showNotification({ type: 'success', message })
+/**
+ * Show message for a success
+ * @param {String} message - Message to show
+ */
+export function showSuccess(message) {
+  return showNotification({ type: 'success', message })
+}
 
 /**
- * @description Dismiss a notification by the given id.
+ * Show message for a error
+ * @param {String} message - Message to show
+ */
+export function showError(message) {
+  return showNotification({ type: 'error', message: `Error: ${message || ''}` })
+}
+
+/**
+ * Dismiss a notification by the given id.
  * @param {Number} id - notification id
  */
-export const dismissNotification = payload => ({
-  type: NOTIFICATION_DISMISS,
-  payload
-})
+export function dismissNotification(payload) {
+  return {
+    type: NOTIFICATION_DISMISS,
+    payload
+  }
+}
 
 /**
- * @description Clear all notifications
+ * Clear all notifications
  */
-export const clearNotifications = () => ({ type: NOTIFICATION_CLEAR })
+export function clearNotifications() {
+  return { type: NOTIFICATION_CLEAR }
+}
