@@ -2,12 +2,12 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { size } from 'lodash'
 import { connect } from 'react-redux'
-import { pure, compose, renderNothing, branch } from 'recompose'
+import { compose, renderNothing, branch } from 'recompose'
 import Snackbar from '@material-ui/core/Snackbar'
 import IconButton from '@material-ui/core/IconButton'
 import CloseIcon from '@material-ui/icons/Close'
-import * as actions from '../actions'
 import { withStyles } from '@material-ui/core/styles'
+import * as actions from '../actions'
 
 const styles = {
   buttonRoot: {
@@ -46,9 +46,13 @@ Notifications.propTypes = {
   dismissNotification: PropTypes.func.isRequired
 }
 
-export default compose(
-  pure,
-  withStyles(styles),
-  connect(({ notifications: { allIds, byId } }) => ({ allIds, byId }), actions),
-  branch(props => !size(props.allIds), renderNothing) // only render if notifications exist
-)(Notifications)
+const enhance = compose(
+  connect(
+    ({ notifications: { allIds, byId } }) => ({ allIds, byId }),
+    actions
+  ),
+  branch(props => !size(props.allIds), renderNothing), // only render if notifications exist
+  withStyles(styles)
+)
+
+export default enhance(Notifications)
