@@ -1,4 +1,17 @@
+import PropTypes from 'prop-types'
+import { compose, setPropTypes } from 'recompose'
 import { reduxForm } from 'redux-form'
-import { ACCOUNT_FORM_NAME } from 'constants'
+import { withStyles } from '@material-ui/core/styles'
+import { ACCOUNT_FORM_NAME } from 'constants/formNames'
+import styles from './AccountForm.styles'
 
-export default reduxForm({ form: ACCOUNT_FORM_NAME })
+export default compose(
+  // set proptypes used in HOCs
+  setPropTypes({
+    onSubmit: PropTypes.func.isRequired // used by reduxForm
+  }),
+  // make the component a redux-form
+  reduxForm({ form: ACCOUNT_FORM_NAME }),
+  // add styles as props
+  withStyles(styles)
+)
