@@ -1,18 +1,18 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Paper from '@material-ui/core/Paper'
-import AccountForm from '../AccountForm'
 import defaultUserImageUrl from 'static/User.png'
-import classes from './AccountPage.scss'
+import AccountForm from '../AccountForm'
 
-export const AccountPage = ({ avatarUrl, updateAccount, profile }) => (
-  <div className={classes.container}>
+export const AccountPage = ({ avatarUrl, updateAccount, profile, classes }) => (
+  <div className={classes.root}>
     <Paper className={classes.pane}>
       <div className={classes.settings}>
-        <div className={classes.avatar}>
+        <div>
           <img
             className={classes.avatarCurrent}
             src={avatarUrl || defaultUserImageUrl}
+            alt=""
           />
         </div>
         <div className={classes.meta}>
@@ -28,6 +28,7 @@ export const AccountPage = ({ avatarUrl, updateAccount, profile }) => (
 )
 
 AccountPage.propTypes = {
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
   avatarUrl: PropTypes.string,
   profile: PropTypes.object,
   updateAccount: PropTypes.func
