@@ -1,7 +1,10 @@
-import { ACCOUNT_PATH as path } from 'constants'
-import component from './components/AccountPage'
+import { Loadable } from 'utils/components'
+import { ACCOUNT_PATH as path } from 'constants/paths'
 
 export default {
   path,
-  component
+  component: Loadable({
+    loader: () =>
+      import(/* webpackChunkName: 'Account' */ './components/AccountPage')
+  })
 }
