@@ -1,87 +1,98 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { connect } from 'react-redux'
-import { compose, withHandlers } from 'recompose'
-import { firebaseConnect, populate } from 'react-redux-firebase'
-import Theme from 'theme'
-import { withNotifications } from 'modules/notification'
-import NewTodoPanel from '../NewTodoPanel'
-import TodosList from '../TodosList'
-import classes from './HomePage.scss'
+import { Link } from 'react-router-dom'
+import {
+  ACCOUNT_PATH,
+  LIST_PATH,
+  LOGIN_PATH,
+  SIGNUP_PATH
+} from 'constants/paths'
 
-// Populate owner from users collection
-const populates = [
-  {
-    child: 'owner', // parameter to populate
-    root: 'users' // collection from which to gather children
-    // childAlias: 'ownerObj' // place result somewhere else on object
-  }
-]
+const authWrapperUrl = 'https://github.com/mjrussell/redux-auth-wrapper'
+const reactRouterUrl = 'https://github.com/ReactTraining/react-router'
 
-// Component enhancer
-const enhance = compose(
-  withNotifications, // adds props.showError from notfication module
-  firebaseConnect([
-    // Create Firebase query for for 20 most recent todos
-    {
-      path: 'todos',
-      queryParams: ['orderByKey', 'limitToLast=10'],
-      populates
-    }
-  ]),
-  // firestoreConnect([{ collection: 'todos' }]) // get data from firestore
-  connect(({ firebase, firebase: { auth } }) => ({
-    uid: auth.uid,
-    todos: populate(firebase, 'todos', populates) // populate todos with users data from redux
-    // todos: firebase.ordered.todos // if using ordering such as orderByChild or orderByKey
-    // todos: firestore.ordered.todos, // firestore data from firestoreConnect
-  })),
-  withHandlers({
-    addNew: props => newTodo =>
-      props.firebase.push('todos', {
-        ...newTodo,
-        owner: props.uid || 'Anonymous'
-      }),
-    onSubmitFail: props => (formErrs, dispatch, err) =>
-      props.showError(formErrs ? 'Form Invalid' : err.message || 'Error')
-  })
-)
-
-const Home = ({ todos, uid, addNew, onSubmitFail }) => (
-  <div
-    className={classes.container}
-    style={{ color: Theme.palette.primary2Color }}>
-    <div className={classes.info}>
-      <span>data loaded from</span>
-      <span>
-        <a href="https://redux-firebasev3.firebaseio.com/">
-          redux-firebasev3.firebaseio.com
-        </a>
-      </span>
-      <span className={classes.note}>
-        <strong>Note: </strong>
-        old data is removed
-      </span>
+export const Home = ({ classes }) => (
+  <div className={classes.root}>
+    <div className="flex-row-center">
+      <h2>Home Route</h2>
     </div>
-    <div className={classes.todos}>
-      <NewTodoPanel
-        disabled={false}
-        onSubmit={addNew}
-        onSubmitFail={onSubmitFail}
-      />
-      <TodosList todos={todos} uid={uid} />
+    <div className="flex-row-center">
+      <div className={classes.section}>
+        <h3>Routing</h3>
+        <span>
+          Redirecting and route protection done using:
+          <div>
+            <span>
+              <a
+                href={reactRouterUrl}
+                target="_blank"
+                rel="noopener noreferrer">
+                react-router
+              </a>
+            </span>
+            <span> and </span>
+            <a href={authWrapperUrl} target="_blank" rel="noopener noreferrer">
+              redux-auth-wrapper
+            </a>
+          </div>
+        </span>
+      </div>
+      <div className={classes.section}>
+        <h4>Logged Out</h4>
+        <span>
+          User is redirected to <pre>/login</pre> if not authenticated and
+          trying to vist:
+        </span>
+        <ul>
+          <li>
+            <Link to={LIST_PATH}>Projects</Link>
+          </li>
+          <li>
+            <Link to={ACCOUNT_PATH}>Account</Link>
+          </li>
+        </ul>
+      </div>
+      <div className={classes.section}>
+        <h4>Logged In</h4>
+        <span>
+          User is redirected to <pre>/projects</pre> if authenticated and trying
+          to vist:
+        </span>
+        <ul>
+          <li>
+            <Link to={LOGIN_PATH}>Login</Link>
+          </li>
+          <li>
+            <Link to={SIGNUP_PATH}>Signup</Link>
+          </li>
+        </ul>
+      </div>
+      <div className={classes.section}>
+        <div>
+          <h4>Forms</h4>
+          <span>Redirecting and route protection done using:</span>
+          <div>
+            <span>
+              <a
+                href={reactRouterUrl}
+                target="_blank"
+                rel="noopener noreferrer">
+                redux-form
+              </a>
+            </span>
+          </div>
+        </div>
+        <span>The following routes use redux-form:</span>
+        <Link to={ACCOUNT_PATH}>
+          <p>Account Page</p>
+        </Link>
+      </div>
     </div>
   </div>
 )
 
-Home.propTypes = {
-  todos: PropTypes.oneOfType([
-    PropTypes.object, // object if using firebase.data
-    PropTypes.array // array if using firebase.ordered
-  ]),
-  addNew: PropTypes.func.isRequired,
-  onSubmitFail: PropTypes.func.isRequired,
-  uid: PropTypes.string
+Home.proptypes = {
+  classes: PropTypes.object.isRequired // from enhancer (withStyles)
 }
 
-export default enhance(Home)
+export default Home
