@@ -1,3 +1,4 @@
 import HomePage from './HomePage'
+import enhance from './HomePage.enhancer'
 
-export default HomePage
+export default enhance(HomePage)
