@@ -1,12 +1,25 @@
-import { withHandlers, pure, compose } from 'recompose'
-import { firebaseConnect } from 'react-redux-firebase'
+import PropTypes from 'prop-types'
+import { withHandlers, compose, setPropTypes } from 'recompose'
+import { withFirebase } from 'react-redux-firebase'
+import { withStyles } from '@material-ui/core/styles'
 import { withNotifications } from 'modules/notification'
 import { UserIsNotAuthenticated } from 'utils/router'
+import styles from './LoginPage.styles'
 
 export default compose(
-  UserIsNotAuthenticated, // redirect to /projects if user is already authed
-  withNotifications, // add props.showError
-  firebaseConnect(), // add props.firebase
+  // redirect to /projects if user is already authed
+  UserIsNotAuthenticated,
+  // add props.showError
+  withNotifications,
+  // add props.firebase
+  withFirebase,
+  // set proptypes used in HOCs
+  setPropTypes({
+    showError: PropTypes.func.isRequired, // used in handlers
+    firebase: PropTypes.shape({
+      login: PropTypes.func.isRequired // used in handlers
+    })
+  }),
   // Handlers as props
   withHandlers({
     onSubmitFail: props => (formErrs, dispatch, err) =>
@@ -18,5 +31,6 @@ export default compose(
     emailLogin: ({ firebase, router, showError }) => creds =>
       firebase.login(creds).catch(err => showError(err.message))
   }),
-  pure // shallow equals comparison on props (prevent unessesary re-renders)
+  // add props.classes
+  withStyles(styles, { withTheme: true })
 )
