@@ -1,18 +1,22 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'react-router'
+import { Link } from 'react-router-dom'
 import GoogleButton from 'react-google-button'
 import Paper from '@material-ui/core/Paper'
-import { SIGNUP_PATH } from 'constants'
+import { SIGNUP_PATH } from 'constants/paths'
 import LoginForm from '../LoginForm'
-import classes from './LoginPage.scss'
 
-export const LoginPage = ({ emailLogin, googleLogin, onSubmitFail }) => (
-  <div className={classes.container}>
+export const LoginPage = ({
+  emailLogin,
+  googleLogin,
+  onSubmitFail,
+  classes
+}) => (
+  <div className={classes.root}>
     <Paper className={classes.panel}>
       <LoginForm onSubmit={emailLogin} onSubmitFail={onSubmitFail} />
     </Paper>
-    <div className={classes.or}>or</div>
+    <div className={classes.orLabel}>or</div>
     <div className={classes.providers}>
       <GoogleButton onClick={googleLogin} />
     </div>
@@ -26,9 +30,10 @@ export const LoginPage = ({ emailLogin, googleLogin, onSubmitFail }) => (
 )
 
 LoginPage.propTypes = {
-  emailLogin: PropTypes.func, // from enhancer (withHandlers)
-  onSubmitFail: PropTypes.func, // from enhancer (withHandlers)
-  googleLogin: PropTypes.func // from enhancer (withHandlers)
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
+  emailLogin: PropTypes.func.isRequired, // from enhancer (withHandlers)
+  onSubmitFail: PropTypes.func.isRequired, // from enhancer (withHandlers)
+  googleLogin: PropTypes.func.isRequired // from enhancer (withHandlers)
 }
 
 export default LoginPage
