@@ -1,5 +1,5 @@
 import { reduxForm } from 'redux-form'
-import { NEW_PROJECT_FORM_NAME } from 'constants'
+import { NEW_PROJECT_FORM_NAME } from 'constants/formNames'
 
 export default reduxForm({
   form: NEW_PROJECT_FORM_NAME,
