@@ -2,19 +2,19 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import ContentAddCircle from '@material-ui/icons/AddCircle'
 import Paper from '@material-ui/core/Paper'
-import classes from './NewProjectTile.scss'
 
 const iconSize = '6rem'
 const iconStyle = { width: iconSize, height: iconSize }
 
-export const NewProjectTile = ({ onClick }) => (
-  <Paper className={classes.container} onClick={onClick}>
+export const NewProjectTile = ({ onClick, classes }) => (
+  <Paper className={classes.root} onClick={onClick}>
     <ContentAddCircle style={iconStyle} />
   </Paper>
 )
 
 NewProjectTile.propTypes = {
-  onClick: PropTypes.func
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
+  onClick: PropTypes.func.isRequired
 }
 
 export default NewProjectTile
