@@ -4,10 +4,15 @@ import Paper from '@material-ui/core/Paper'
 import IconButton from '@material-ui/core/IconButton'
 import Tooltip from '@material-ui/core/Tooltip'
 import DeleteIcon from '@material-ui/icons/Delete'
-import classes from './ProjectTile.scss'
 
-export const ProjectTile = ({ name, onSelect, onDelete, showDelete }) => (
-  <Paper className={classes.container}>
+export const ProjectTile = ({
+  name,
+  onSelect,
+  onDelete,
+  showDelete,
+  classes
+}) => (
+  <Paper className={classes.root}>
     <div className={classes.top}>
       <span className={classes.name} onClick={onSelect}>
         {name || 'No Name'}
@@ -24,6 +29,7 @@ export const ProjectTile = ({ name, onSelect, onDelete, showDelete }) => (
 )
 
 ProjectTile.propTypes = {
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
   name: PropTypes.string,
   onSelect: PropTypes.func.isRequired,
   onDelete: PropTypes.func,
