@@ -1,13 +1,22 @@
-import React, { cloneElement } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 import { isEmpty } from 'react-redux-firebase'
+import { Route, Switch } from 'react-router-dom'
+import ProjectRoute from 'routes/Projects/routes/Project'
 import ProjectTile from '../ProjectTile'
 import NewProjectTile from '../NewProjectTile'
 import NewProjectDialog from '../NewProjectDialog'
-import classes from './ProjectsPage.scss'
+
+const renderChildren = (routes, match, parentProps) =>
+  routes.map(route => (
+    <Route
+      key={`${match.url}-${route.path}`}
+      path={`${match.url}/${route.path}`}
+      render={props => <route.component {...parentProps} {...props} />}
+    />
+  ))
 
 export const ProjectsPage = ({
-  children,
   projects,
   collabProjects,
   auth,
@@ -15,34 +24,45 @@ export const ProjectsPage = ({
   toggleDialog,
   deleteProject,
   addProject,
+  classes,
+  match,
   goToProject
-}) =>
-  children ? (
-    cloneElement(children, { auth })
-  ) : (
-    <div className={classes.container}>
-      <NewProjectDialog
-        onSubmit={addProject}
-        open={newDialogOpen}
-        onRequestClose={toggleDialog}
-      />
-      <div className={classes.tiles}>
-        <NewProjectTile onClick={toggleDialog} />
-        {!isEmpty(projects) &&
-          projects.map((project, ind) => (
-            <ProjectTile
-              key={`Project-${project.key}-${ind}`}
-              name={project.value.name}
-              onSelect={() => goToProject(project.key)}
-              onDelete={() => deleteProject(project.key)}
-            />
-          ))}
-      </div>
-    </div>
-  )
+}) => (
+  <Switch>
+    {/* Child routes */}
+    {renderChildren([ProjectRoute], match, { auth })}
+    {/* Main Route */}
+    <Route
+      exact
+      path={match.path}
+      render={() => (
+        <div className={classes.root}>
+          <NewProjectDialog
+            onSubmit={addProject}
+            open={newDialogOpen}
+            onRequestClose={toggleDialog}
+          />
+          <div className={classes.tiles}>
+            <NewProjectTile onClick={toggleDialog} />
+            {!isEmpty(projects) &&
+              projects.map((project, ind) => (
+                <ProjectTile
+                  key={`Project-${project.id}-${ind}`}
+                  name={project.name}
+                  onSelect={() => goToProject(project.id)}
+                  onDelete={() => deleteProject(project.id)}
+                />
+              ))}
+          </div>
+        </div>
+      )}
+    />
+  </Switch>
+)
 
 ProjectsPage.propTypes = {
-  children: PropTypes.object, // from react-router
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
+  match: PropTypes.object.isRequired, // from enhancer (withRouter)
   auth: PropTypes.object, // from enhancer (connect + firebaseConnect - firebase)
   projects: PropTypes.array, // from enhancer (connect + firebaseConnect - firebase)
   newDialogOpen: PropTypes.bool, // from enhancer (withStateHandlers)
