@@ -1,16 +1,36 @@
+import PropTypes from 'prop-types'
 import { compose } from 'redux'
 import { connect } from 'react-redux'
-import { firebaseConnect, getVal } from 'react-redux-firebase'
+import { get } from 'lodash'
+import { firebaseConnect } from 'react-redux-firebase'
+import { withStyles } from '@material-ui/core/styles'
+import { withRouter } from 'react-router-dom'
+import { setPropTypes, withProps } from 'recompose'
 import { spinnerWhileLoading } from 'utils/components'
 import { UserIsAuthenticated } from 'utils/router'
+import styles from './ProjectPage.styles'
 
 export default compose(
-  // redirect to /login if user is not logged in
+  // Redirect to /login if user is not logged in
   UserIsAuthenticated,
-  firebaseConnect(({ params }) => [{ path: `projects/${params.projectname}` }]),
-  connect(({ firebase: { data } }, { params }) => ({
-    project: getVal(data, `projects/${params.projectname}`)
+  // Add props.match
+  withRouter,
+  // Set proptypes of props used in HOCs
+  setPropTypes({
+    match: PropTypes.shape({
+      params: PropTypes.shape({
+        projectId: PropTypes.string.isRequired
+      }).isRequired
+    }).isRequired
+  }),
+  withProps(({ match: { params: { projectId } } }) => ({
+    projectId
+  })),
+  firebaseConnect(({ projectId }) => [{ path: `projects/${projectId}` }]),
+  connect(({ firebase: { data } }, { projectId }) => ({
+    project: get(data, `projects.${projectId}`)
   })),
   // Show loading spinner while project is loading
-  spinnerWhileLoading(['project'])
+  spinnerWhileLoading(['project']),
+  withStyles(styles)
 )
