@@ -1,13 +1,18 @@
 import { withFirebase } from 'react-redux-firebase'
-import { withHandlers, pure, compose } from 'recompose'
+import { withHandlers, compose } from 'recompose'
 import { UserIsNotAuthenticated } from 'utils/router'
 import { withNotifications } from 'modules/notification'
+import { withStyles } from '@material-ui/core/styles'
+import styles from './SignupPage.styles'
 
 export default compose(
-  UserIsNotAuthenticated, // redirect to list page if logged in
-  withNotifications, // add props.showError
-  withFirebase, // add props.firebase (firebaseConnect() can also be used)
-  // Handlers
+  // Redirect to list page if logged in
+  UserIsNotAuthenticated,
+  // Add props.showError
+  withNotifications,
+  // Add props.firebase (used in handlers)
+  withFirebase,
+  // Add handlers as props
   withHandlers({
     onSubmitFail: props => (formErrs, dispatch, err) =>
       props.showError(formErrs ? 'Form Invalid' : err.message || 'Error'),
@@ -23,5 +28,6 @@ export default compose(
         })
         .catch(err => showError(err.message))
   }),
-  pure // shallow equals comparison on props (prevent unessesary re-renders)
+  // Add styles as props.classes
+  withStyles(styles)
 )
