@@ -1,19 +1,17 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { Link } from 'react-router'
+import { Link } from 'react-router-dom'
 import GoogleButton from 'react-google-button'
 import Paper from '@material-ui/core/Paper'
-import { LOGIN_PATH } from 'constants'
+import { LOGIN_PATH } from 'constants/paths'
 import SignupForm from '../SignupForm'
 
-import classes from './SignupPage.scss'
-
-export const SignupPage = ({ emailSignup, googleLogin, onSubmitFail }) => (
-  <div className={classes.container}>
+const SignupPage = ({ emailSignup, googleLogin, onSubmitFail, classes }) => (
+  <div className={classes.root}>
     <Paper className={classes.panel}>
       <SignupForm onSubmit={emailSignup} onSubmitFail={onSubmitFail} />
     </Paper>
-    <div className={classes.or}>or</div>
+    <div className={classes.orLabel}>or</div>
     <div className={classes.providers}>
       <GoogleButton onClick={googleLogin} />
     </div>
@@ -27,9 +25,10 @@ export const SignupPage = ({ emailSignup, googleLogin, onSubmitFail }) => (
 )
 
 SignupPage.propTypes = {
-  emailSignup: PropTypes.func, // from enhancer (withHandlers - firebase)
-  googleLogin: PropTypes.func, // from enhancer (withHandlers - firebase)
-  onSubmitFail: PropTypes.func // from enhancer (reduxForm)
+  classes: PropTypes.object.isRequired, // from enhancer (withStyles)
+  emailSignup: PropTypes.func.isRequired, // from enhancer (withHandlers)
+  googleLogin: PropTypes.func.isRequired, // from enhancer (withHandlers)
+  onSubmitFail: PropTypes.func.isRequired // from enhancer (reduxForm)
 }
 
 export default SignupPage
