@@ -1,7 +1,10 @@
-import { SIGNUP_PATH as path } from 'constants'
-import component from './components/SignupPage'
+import { Loadable } from 'utils/components'
+import { SIGNUP_PATH as path } from 'constants/paths'
 
 export default {
   path,
-  component
+  component: Loadable({
+    loader: () =>
+      import(/* webpackChunkName: 'Signup' */ './components/SignupPage')
+  })
 }
