@@ -1,26 +1,33 @@
-import { UserAuthWrapper } from 'redux-auth-wrapper'
-import { browserHistory } from 'react-router'
-import { LIST_PATH } from 'constants'
+import { connectedRouterRedirect } from 'redux-auth-wrapper/history4/redirect'
+import locationHelperBuilder from 'redux-auth-wrapper/history4/locationHelper'
+import createHistory from 'history/createBrowserHistory'
 import LoadingSpinner from 'components/LoadingSpinner'
+import { LIST_PATH } from 'constants/paths'
+
+const locationHelper = locationHelperBuilder({})
+const history = createHistory()
 
 const AUTHED_REDIRECT = 'AUTHED_REDIRECT'
 const UNAUTHED_REDIRECT = 'UNAUTHED_REDIRECT'
 
 /**
- * @description Higher Order Component that redirects to `/login` instead
+ * Higher Order Component that redirects to `/login` instead
  * rendering if user is not authenticated (default of redux-auth-wrapper).
  * @param {Component} componentToWrap - Component to wrap
  * @return {Component} wrappedComponent
  */
-export const UserIsAuthenticated = UserAuthWrapper({
+export const UserIsAuthenticated = connectedRouterRedirect({
+  redirectPath: '/login',
+  AuthenticatingComponent: LoadingSpinner,
   wrapperDisplayName: 'UserIsAuthenticated',
-  LoadingComponent: LoadingSpinner,
-  authSelector: ({ firebase: { auth } }) => auth,
+  // Want to redirect the user when they are done loading and authenticated
+  authenticatedSelector: ({ firebase: { auth } }) =>
+    !auth.isEmpty && !!auth.uid,
   authenticatingSelector: ({ firebase: { auth, isInitializing } }) =>
     !auth.isLoaded || isInitializing,
-  predicate: auth => !auth.isEmpty,
   redirectAction: newLoc => dispatch => {
-    browserHistory.replace(newLoc)
+    // Use push, replace, and go to navigate around.
+    history.push(newLoc)
     dispatch({
       type: UNAUTHED_REDIRECT,
       payload: { message: 'User is not authenticated.' }
@@ -29,31 +36,29 @@ export const UserIsAuthenticated = UserAuthWrapper({
 })
 
 /**
- * @description Higher Order Component that redirects to listings page or most
+ * Higher Order Component that redirects to listings page or most
  * recent route instead rendering if user is not authenticated. This is useful
  * routes that should not be displayed if a user is logged in, such as the
  * login route.
  * @param {Component} componentToWrap - Component to wrap
  * @return {Component} wrappedComponent
  */
-export const UserIsNotAuthenticated = UserAuthWrapper({
+export const UserIsNotAuthenticated = connectedRouterRedirect({
+  AuthenticatingComponent: LoadingSpinner,
   wrapperDisplayName: 'UserIsNotAuthenticated',
   allowRedirectBack: false,
-  LoadingComponent: LoadingSpinner,
-  failureRedirectPath: (state, props) =>
-    // redirect to page user was on or to list path
-    props.location.query.redirect || LIST_PATH,
-  authSelector: ({ firebase: { auth } }) => auth,
+  // Want to redirect the user when they are done loading and authenticated
+  authenticatedSelector: ({ firebase: { auth } }) => auth.isEmpty,
   authenticatingSelector: ({ firebase: { auth, isInitializing } }) =>
     !auth.isLoaded || isInitializing,
-  predicate: auth => auth.isEmpty,
+  redirectPath: (state, ownProps) =>
+    locationHelper.getRedirectQueryParam(ownProps) || LIST_PATH,
   redirectAction: newLoc => dispatch => {
-    browserHistory.replace(newLoc)
-    dispatch({ type: AUTHED_REDIRECT })
+    // Use push, replace, and go to navigate around.
+    history.push(newLoc)
+    dispatch({
+      type: AUTHED_REDIRECT,
+      payload: { message: 'User is not authenticated.' }
+    })
   }
 })
-
-export default {
-  UserIsAuthenticated,
-  UserIsNotAuthenticated
-}
